Fix build with askama 0.16 and rusqlite 0.40
- askama 0.16 treats `set` specially in if conditions (parses `x == set`
  as a let-condition chain); rename the repo-set loop variable to `rs`/`name`
- rusqlite 0.40 no longer implements ToSql for u64; cast issue id/number to
  i64 at the DB boundary

+++ src/db.rs
@@ -391,8 +391,8 @@ CREATE TABLE IF NOT EXISTS org_repos (
                created_at = COALESCE(issues.created_at, excluded.created_at)",
             params![
                 repo_id,
-                issue.id,
-                issue.number,
+                issue.id as i64,
+                issue.number as i64,
                 kind,
                 issue.title,
                 issue.state,
